% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STIP.R
\name{STIP}
\alias{STIP}
\title{Perform STIP}
\usage{
STIP(fit, gl)
}
\arguments{
\item{fit}{a numeric matrix of gene expression values. Each row is a gene and each column is a cell. Cells are ordered according to pseudotime.}

\item{gl}{a character vector of gene names that will be highlighted.}
}
\value{
A ggplot object.
}
\description{
State Transition Inference Prediction
}
\details{
This function generates a plot that performs (STIP) State Transition Inference Prediction.
}
\examples{
set.seed(12345)
fit <- t(apply(matrix(rnorm(1000),nrow=10),1,sort))
row.names(fit) <- LETTERS[1:10]
gl <- c("C", "G", "E", "I", "A")
STIP(fit,gl)
}
\author{
Zhicheng Ji, Zeyu Chen <zji4@zji4.edu>
}
